requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
blinker==1.4
certifi==2024.8.30
cloudinary==1.36.0
colorama==0.4.6
cryptography==3.4.8
dbus-python==1.2.18
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.114.2
greenlet==3.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.20.2
Hypercorn==0.17.3
hyperframe==6.0.1
icecream==2.1.3
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
mysqlclient==2.1.1
netifaces==0.11.0
oauthlib==3.2.0
passlib==1.7.4
priority==2.0.0
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
PyMySQL==1.1.1
pyparsing==2.4.7# Step 1: Install build tools
sudo apt-get update
sudo apt-get install -y build-essential python3-dev

# Step 2: Install PyYAML using a pre-built wheel
pip install PyYAML==5.4.1
python-dotenv==1.0.1
python-multipart==0.0.9
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
systemd-python==234
taskgroup==0.0.0a4
tomli==2.0.1
typing_extensions==4.12.2
